Switch to proxy since that's what Next 16 uses instead of middleware
